requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
pytest
streamlit
python-multipart


